Datatyp changes (test functions)
Changed datatypes for the calc and conv functions

+++ global.c
@@ -1,11 +1,11 @@
 #include "global.h"
 
 //___GLOBAL VARIABLES___
-volatile float vref_h;      	// Vref+ for ADC10 (calibrated sensor)
-volatile float *ptr_vref_h; 	// Vref+ as raw
+volatile int vref_h;      	// Vref+ for ADC10 (calibrated sensor)
+volatile int *ptr_vref_h; 	// Vref+ as raw
     
-volatile float vref_l;       	// Vref- for ADC10 (calibrated sensor)
-volatile float *ptr_vref_l;		// Vref- as raw
+volatile int vref_l;       	// Vref- for ADC10 (calibrated sensor)
+volatile int *ptr_vref_l;		// Vref- as raw
 
-volatile float vref_vcc;        // VCC at Vref
-volatile float *ptr_vref_vcc;	// Vcc as raw
+volatile int vref_vcc;        // VCC at Vref
+volatile int *ptr_vref_vcc;	// Vcc as raw
